requirements/base.txt
Django==1.10.1
psycopg2==2.6
djangorestframework==3.4.7
celery==3.1.23
sqlalchemy==1.0.13
PyYAML==3.11
django-rest-swagger==2.0.5
markdown==2.6.7

